4 transaction check

+++ ProSoft.UI/Areas/Treasury/Controllers/AccSafeCashController.cs
@@ -11,6 +11,8 @@ using ProSoft.EF.Models.Medical.HospitalPatData;
 using ProSoft.EF.Models.Treasury;
 using ArabicNumbersConverter;
 using Humanizer;
+using System.Globalization;
+using System.Security.Policy;
 
 namespace ProSoft.UI.Areas.Treasury.Controllers
 {
@@ -26,7 +28,8 @@ namespace ProSoft.UI.Areas.Treasury.Controllers
             _accSafeCashRepo = accSafeCashRepo;
             _userCashNoRepo = userCashNoRepo;
         }
-        public async Task<IActionResult> Index(string docType,string? flagType)
+
+        public async Task<IActionResult> Index(string docType,string? flagType ,string? errorMessage)
         {
             var userCode = int.Parse(User.Claims.FirstOrDefault(c => c.Type == "User_Code").Value);
             var userCashNoDTO = await _userCashNoRepo.GetSafeTransByIdAsync(userCode);
@@ -36,12 +39,41 @@ namespace ProSoft.UI.Areas.Treasury.Controllers
 
             List<AccSafeCashViewDTO> accSafeCashs = await _accSafeCashRepo
             .GetAccSafeCashAsync(docType, flagType, fYear, safeCode);
+
+            // Convert the numbers to Arabic words
+            foreach (var cash in accSafeCashs)
+            {
+                if (cash.ValuePay !=null)
+                {
+                    cash.ValuePayWord = ConvertNumberToArabicWords(decimal.Parse(cash.ValuePay.ToString()));
+                }
+            }
             ViewBag.docType = docType;
             ViewBag.safeCode = safeCode;
             ViewBag.fYear = fYear;
-
+            ViewBag.error = errorMessage;
             return View(accSafeCashs);
         }
+
+        //// Method to convert numbers to Arabic words
+        private string ConvertNumberToArabicWords(decimal number)
+        {
+            long wholePart = (long)number;
+            int fractionalPart = (int)((number - wholePart) * 100); // Adjust this if you need more precision
+
+            string wholePartInWords = wholePart.ToWords(new CultureInfo("ar"));
+            string fractionalPartInWords = fractionalPart.ToWords(new CultureInfo("ar"));
+
+            if (fractionalPart > 0)
+            {
+                return $"{wholePartInWords} و {fractionalPartInWords}";
+            }
+            else
+            {
+                return wholePartInWords;
+            }
+        }
+
         //Ajax In Add_StockTrans
         public async Task<IActionResult> GetSubCodesFromAcc(string id)
         {
@@ -99,6 +131,14 @@ namespace ProSoft.UI.Areas.Treasury.Controllers
         public async Task<IActionResult> Delete_PaymentReceipt(int id)
         {
             AccSafeCash accSafeCash = await _accSafeCashRepo.GetByIdAsync(id);
+
+            bool hasRelatedData = await _accSafeCashRepo.HasRelatedDataAsync(id);
+            if (hasRelatedData)
+            {
+                var errorMessage = "Cannot delete this record because it has related data in another table!";
+                //ViewBag.ErrorMessage = "Cannot delete this record because it has related data in another table.";
+                return RedirectToAction("Index", "AccSafeCash", new { docType = "SFCIN", flagType = "oneANDtwo" , errorMessage });
+            }
 
             await _accSafeCashRepo.DeleteAsync(accSafeCash);
             await _accSafeCashRepo.SaveChangesAsync();
