+++ ProSoft.UI/Areas/Treasury/Controllers/TransferPermissionController.cs
@@ -1,9 +1,11 @@
-﻿using Microsoft.AspNetCore.Authorization;
+﻿using Humanizer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using ProSoft.EF.DTOs.Accounts;
 using ProSoft.EF.DTOs.Treasury;
 using ProSoft.EF.IRepositories.Treasury;
 using ProSoft.EF.Models.Treasury;
+using System.Globalization;
 
 namespace ProSoft.UI.Areas.Treasury.Controllers
 {
@@ -18,7 +20,7 @@ namespace ProSoft.UI.Areas.Treasury.Controllers
             _accSafeCashRepo = accSafeCashRepo;
             _userCashNoRepo = userCashNoRepo;
         }
-        public async Task<IActionResult> Index(string docType, string? flagType)
+        public async Task<IActionResult> Index(string docType, string? flagType, string? errorMessage)
         {
             var userCode = int.Parse(User.Claims.FirstOrDefault(c => c.Type == "User_Code").Value);
             var userCashNoDTO = await _userCashNoRepo.GetSafeTransByIdAsync(userCode);
@@ -28,12 +30,43 @@ namespace ProSoft.UI.Areas.Treasury.Controllers
 
             List<AccSafeCashViewDTO> accSafeCashs = await _accSafeCashRepo
             .GetAccSafeCashAsync(docType, flagType, fYear, safeCode);
+
+            // Convert the numbers to Arabic words
+            foreach (var cash in accSafeCashs)
+            {
+                if (cash.ValuePay != null)
+                {
+                    cash.ValuePayWord = ConvertNumberToArabicWords(decimal.Parse(cash.ValuePay.ToString()));
+                }
+            }
+
             ViewBag.docType = docType;
             ViewBag.safeCode = safeCode;
             ViewBag.fYear = fYear;
+            ViewBag.error = errorMessage;
 
             return View(accSafeCashs);
         }
+
+        //// Method to convert numbers to Arabic words
+        private string ConvertNumberToArabicWords(decimal number)
+        {
+            long wholePart = (long)number;
+            int fractionalPart = (int)((number - wholePart) * 100); // Adjust this if you need more precision
+
+            string wholePartInWords = wholePart.ToWords(new CultureInfo("ar"));
+            string fractionalPartInWords = fractionalPart.ToWords(new CultureInfo("ar"));
+
+            if (fractionalPart > 0)
+            {
+                return $"{wholePartInWords} و {fractionalPartInWords}";
+            }
+            else
+            {
+                return wholePartInWords;
+            }
+        }
+
         //Ajax In Add_StockTrans
         public async Task<IActionResult> GetSubCodesFromAcc(string id)
         {
@@ -92,6 +125,13 @@ namespace ProSoft.UI.Areas.Treasury.Controllers
         {
             AccSafeCash accSafeCash = await _accSafeCashRepo.GetByIdAsync(id);
 
+            bool hasRelatedData = await _accSafeCashRepo.HasRelatedDataAsync(id);
+            if (hasRelatedData)
+            {
+                var errorMessage = "Cannot delete this record because it has related data in another table!";
+                //ViewBag.ErrorMessage = "Cannot delete this record because it has related data in another table.";
+                return RedirectToAction("Index", "TransferPermission", new { docType = "SFTOT", flagType = "oneANDtwoAndthree", errorMessage });
+            }
             await _accSafeCashRepo.DeleteAsync(accSafeCash);
             await _accSafeCashRepo.SaveChangesAsync();
             return RedirectToAction("Index", "TransferPermission", new { docType = "SFTOT", flagType = "oneANDtwoAndthree" });
